Remove channel buffers from Write path
This makes a Write report as completing when it can't possibly have.
It's still not totally correct in this case, but it's far more likely to
do the right thing when a Write that sends a packet to a channel
requires the other end to be ready to receive and process that Write.

Close #3